Memory pool allocation + more
This commit introduces fmempool, a pool-based memory allocator that can function like the `malloc` family of function in libc. It does incur significantly more overhead due to the ability to free regions without providing the size (and as such, the order needs to stored elsewhere). However, this is not too terrible (6.25% of the region size needs to be provided as additional bookkeeping space).

These new functions and the fpage function have been marked with warn-unused-result.

`ferr_invalid_parameter` has been renamed to `ferr_invalid_argument`, to more accurately reflect the issue. Parameters are what the function defines as expectations, but arguments are what callers provide for parameters.

Also, fix page allocation: we were allocating less than what was needed because we were determining order incorrectly.