Initial x86_64 interrupts code
Now that we have interrupts and have the potential for functions to be called from ISRs, we need some IRQ-safe locks (and locks in general).

This is necessary e.g. for memory allocation: we don't want to be in middle of an allocation and suddenly receive an interrupt which also needs to allocate memory, because then the memory bookkeeping could be in an invalid state when the second interrupt occurs.

Soon, we'll also need per-CPU data, but that's less of a pressing concern.