Implement spinlocks and per-CPU data on x86_64
Interrupt-safe spinlocks are very important so that we can allocate memory safely without the fear of entering an interrupt in the middle of an allocation and having the memory state corrupted.

Per-CPU data is necessary to keep track of the outstanding-interrupt-disable count for each CPU. For now, since we only have one CPU, it's just a pointer to a static variable. But the current design allows it to be easily extended when we enable SMP.

This commit also goes ahead and adds locks to all the memory allocation systems.