Clean up the source some more
All of the current code was actually written in 2021, so all the license headers should say that instead of 2020.

Also, add descriptions for files that were missing them. And the descriptions should contain the full filename, starting from the subproject root.

Finally, all of the architecture-specific definitions should be prefixed with `farch`.