Threads & more!
This commit introduces kernel threads, a basic scheduler, and plenty of fixes for a bunch of subsystems.

## New code

The scheduler and threading subsystems are brand new subsystems that do exactly what they sound like.

The threading subsystem manages information and saved context about threads. Each thread has its own manager, which can be any manager that implements the necessary functions. In the code added now, the scheduler is the only thread manager available.

In addition to simply keeping track of thread data/contex, the threading subsystem also provides an API for users to manage thread state. Behind the scenes, this API asks the thread's manager to perform various operations, such as suspension, resumption,  and killing.

Threads are reference counted, as there could be many different pieces of code that need access to a thread and knowing who's resposible for freeing the memory would be a PITA otherwise (e.g. the scheduler might own it for a time, but whoever gave it to the scheduler might want to keep it around after it dies and check the exit data left by the thread).

## Fixes/changes to existing code

First, an important build/source change: assembly can now be preprocessed (with the `S` extension; the capitalization is important). Additionally, using Clang, we can calculate the offsets of structure members, allowing us to easily use those structures in assembly.

The interrupts subsystem now saves and loads all registers on interrupt enter and exit. This is done with a separate assembly file filled with interrupt wrappers that perform this task.

The console subsystem now locks the logging lock for the entire duration of a log message (rather than a single character as it was before). This prevents jumbled messages from appearing in the console. Ideally, the console should instead save messages to a buffer and have a background thread output them to the screen later to prevent the operation from blocking preemption for too long.

In the mempool subsystem, `fmempool_reallocate` has been fixed in the expansion cases; I had forgotten to write the output address for those cases.

In the paging subsystem, temporary mapping is instead replaced with fixed offset mapping (for physical memory access).

In the timers subsystem, timers are now no longer fired one after the other if more than one is ready. Instead, each is scheduled after the next, with a minimum delay of 1 microsecond between fires. Additionally, a few silly mistakes in the code have been fixed.

Finally, the memory map created by bootstrapers and passed to the kernel on entry now requires an entry for the kernel stack, to avoid it being treated as general memory by the paging subsystem and being allocated for other uses (thereby corrupting the stack).

## TODO

ARM support for threads!