waitqs, semaphores, and workers, oh my!
As the brief commit description says, this commit adds support for waitqs, semaphore, and workers.

Waitqs and conditional thread sleeping are the most important new features. Together, these two features allow threads to suspend themselves until a condition is met. Additionally, timed thread sleeping has also been added, which allows threads to sleep for a certain amount of time (TODO: there should be an easy way to wait for a waitq but also have a timeout in case the waitq doesn't wake us up in the given amount of time).

A minor addition is that we now have a Doxyfile for generating documentation, but this a work-in-progress (it doesn't generate it as I would like).

One thing to note: all these changes work on x86_64 but not AARCH64. That's the next task: getting them to work on AARCH64.