Add reboot and basic UBSan support
Also, rebuilding is now disabled by default when running the test script. To build and then run QEMU using only the test script, you must give it the `-b`/`--build` option.

And, AARCH64 is broken. Yet again. This time with regards to UBSan. It hangs after "entering threaded kernel startup" and LLDB says it's stuck in the interrupt handlers for lower ELs? Which is... weird?

Unfortunately, both x86_64 and AARCH64 are still broken with the workers :(. This is dumb :((((