A *mountain* of changes; check the description
Sorry I wasn't able to split this commit up; a lot of these changes depend on each other to work properly.

This commit introduces userspace! We can now load processes and run code in userspace. There's also a syscall interface that allows userspace to call into the kernel and perform certain actions like I/O and memory allocation.

Work has already been on a dynamic linker for dynamic userspace binaries so that code can be split up into libraries.

libsys is going to be the base library for userspace; it will provide a stable ABI for userspace to interface with the kernel. libc will depend on it and use it to implement the standard C library.

I've already made a very important decision regarding libsys. Unlike the UNIX philosophy where everything is a file, Anillo will follow a more object-based approach. So far, files and streams are both objects.

And that's another thing: libsys differentiates between files and streams. Files can be randomly accessed, whereas streams cannot. Currently, the standard console output is classified as a stream while the process binary is classified as a file (for obvious reasons). In the future, network connections like TCP sockets will be classified as streams, since they cannot be read from or written to at arbitrary locations.

I've created a CMake script for adding files created during the build to the ramdisk. Currently, it only supports adding targets, but it can be extended to support adding arbitrary files.

I'm sure there are many more things I missed; please check the diff.