A slew of changes; check the description
So we've switched to Mach-O binaries because of their more sophisticated dynamic linking (the two-level namespace and proper reexporting support is what really sells it).

Also we have a dynamic linker now. It works, too. Work has started on a libc implementation as well. The kernel can now do some basic stack tracing. A couple of fixes in the kernel have also fixed baremetal booting.

So yeah... lots of stuff has changed. Check the diff.