Add sys_event in libsys and futex_associate syscall
futex_associate is used to have the kernel wakeup a futex when certain kernel events occur. Currently, thread death is the only event that can be associated with a futex.

However, something is currently broken, because when deallocating the stack for the primary thread in sysman, the locking subsystem panics while unlocking a lock for a virtual address space region.