ferro: Fix userspace thread creation wrt process references
We weren't retaining the process for additional threads created via the
thread_create syscall.

This could've caused many different issues; interestingly, it manifested
itself as a bug with using an invalid address for the lock of the regions
in the address space of the process, because the memory for the address
space was being released when the process was released early.