libsys: Add an API for using channels
This API is meant to be the foundation for IPC in Anillo. It is meant to be
a general and minimal interface for the functionality provided by the kernel.
While usable, it is generally too low-level to be used comfortably (especially asynchronously).
A future commit will introduce a library providing a much simpler interface
for using channels that enables them to be used
asynchronously in a comfortable way.