libsys: Add an API for monitors
Monitors can be used to monitor libsys objects and wait for certain events
to be triggered on those objects.

Like with the channel API, while this API *can* be used as-is, it is
generally too low-level to be used comfortably.
A future commit will introduce a library that can monitor
libsys objects for events and manage monitors internally, providing a simple API.