Add libeve, an event loop library
This library builds on the interface provided by libsys to provide a
simple event loop API.

This library can manage sending and receiving messages asynchronously on channels
and scheduling work items to run on the event loop.
Additionally, it provides pseudo-synchronous APIs that allow event loop work
to perform blocking calls (like sending a message or waiting for a semaphore)
without blocking the event loop.