ferro: Allow threads to be blocked
Blocking is a separate state from suspension. Suspension indicates "wake
me up later when something important happens" but blocking indicates "I
cannot run *at all* until something else happens". In other words,
suspension is like a weaker form of blocking; you're allowed to be woken
spuriously from a suspension, but you're *not* allowed to be woken
spuriously from a block.

The main use case for blocking threads is preventing them from running
while another thread modifies or observes their execution in some way
(e.g. a signal handler for uthreads).

Another important distinction between suspension and blocking:
even if a thread is suspended 5 times, it only needs to be resumed once
for it to start running again; if a thread is blocked 5 times, however,
it needs to be unblocked 5 times for it to start running again.

Blocked threads can still be suspended, resumed, or killed; the only
thing that blocking prevents is being scheduled to run on a CPU.