ferro: Implement support for signals
Signals allow threads to receive asynchronous notifications about events.

It is *not* recommended to use signals for IPC or even for inter-thread
communication; however, there are certain events for which there is no
other suitable way to handle them (e.g. page faults).

Signals are also required for full implementations of the C library and
POSIX standard.