Some updates to how signals are implemented
After reviewing some of the features of POSIX signals, it turns out that
we actually need dispatched signals to be contained entirely within userspace.
In other words, the only thing that can be in-kernel is the signal queue;
when we finally setup a signal to be handled, all the information necessary
to jump back after the signal handler is done needs to be in userspace.
This is because POSIX allows you to jump away during signal handlers and
continue executing as if nothing had happened.