Add support for special signals (e.g. page faults)
These special signals *must* be handled, otherwise they kill the process.
Unlike most POSIX systems, it is allowed for the signals to be handled
by another thread; as long as it is handling the signal, the original
thread is blocked from executing.

Additionally, this commit also adds a syscall for fetching constants
like the page size or thread context size to avoid hard-coding those values
in userspace.