kernel: Initial x86_64 SMP code
The APs (application processors) currently do nothing besides spin with interrupts halted,
but they are successfully started!

Next comes the hard part: updating other subsystems
to properly support SMP operation.
You might think that the scheduler would be the hardest,
but actually that has already been designed with SMP
(at least partially) in mind.
In reality, the paging subsystem is probably going to be the trickiest.