Reduce unnecessary interrupts in netman
This is accomplished by reading from the ICR register in the interrupt handler
in the kernel, using a new pciman API for reading/writing values to
a BAR region on an interrupt.

This is necessary because the E1000 only clears the interrupt-enable mask
when reading from the ICR register, so this must be done in the interrupt
handler in the kernel to avoid further interrupts.

Because we want to minimize driver code in kernel-space (that is,
after all, the entire point of a microkernel), we provide an API for drivers
to ask pciman to perform a read and/or write in the interrupt handler.