Build static versions of libeve and libspooky
We're going to need these later, as we're going to move to having all
binaries in the ramdisk be static binaries so the kernel can load them
directly without a dynamic linker. This is necessary because we're going
to eliminate VFS syscalls entirely; we're going to have the entire VFS
system (including FS drivers) implemented in userspace. Thus, we cannot
have the dynamic linker for these early-boot managers, so we need to
make them static binaries.

The plan is to have the kernel start up the VFS manager and ramdisk FS driver
and then the system manager, once the VFS services are loaded and available.
The system manager can then load up other FS drivers as necessary.