kernel: Fix another race condition in the core channel code
The code in `kernel/src/core/channels.c` contains a comment explaining
why this change is necessary, but I'd like to dive into how I was able
to notice and debug this.

Despite being a "race" condition, this bug ironically only popped up on
slow enough systems; I was only able to consistently reproduce it on
a laptop running in QEMU with TCG emulation.

What I observed was that sometimes userspace would send a message on a
channel that was being monitored, the monitor would be triggered, and
the libeve monitor thread would be awoken instantly, which would then
try to receive the message. However, the original thread that sent the
message had still not incremented the message-removal semaphore, so the
message receive would fail with `ferr_no_wait` and userspace would
consider that event handled, leaving a message in the channel unhandled
forever (since the sender would be waiting for a reply, so no more
events would ever come in).