Add an API to attach data to messages; add shared data
The goal of these additions is to reduce the number of copies of data
required to transmit it over IPC.

For non-shared data, the number of copies remains the same. However,
for shared data, two copies have been eliminated (one on sending and
another on receiving). The advantage of supporting both is that we can
seamlessly use shared data objects just like non-shared data objects and
vice versa.

This commit also updates libspooky to use this new API to attach data
arguments directly to the invocation message. This means that, if the
caller/callee provides a shared data object as an argument, it will be
attached to the message as such and will avoid additional copies.

This commit does not introduce an API to actually create shared data
objects, however (simply because I don't have time to do that right now).