Significantly improve image load times
Because we now handle VFS in userspace, file reads are more expensive.
The size of the read doesn't generally matter too much (although the
last few commits have optimized large reads to avoid copying as much).
Rather, the bottleneck is the number of reads, since these are now IPC
calls and the round-trip is significantly slower than what it used to be.

Thus, we now read load commands all at once instead of sprinkling lots
of reads as we go through the image. This is significantly faster and
we should've been doing this from the beginning anyways.