[syscalls] Use physical addresses for futexes
This allows futexes to work as expected with shared memory.

TODO: we should try to trigger a page fault on the address to have it bound
by the paging subsystem in case this is a valid user address
that simply hasn't been accessed yet. In the typical futex usecase, this
shouldn't occur (userspace is supposed to check the futex before making
syscalls to wait on it), but it *should* be supported.