Major refactor: get rid of server channels; launch sysman directly
This commit gets rid of dedicated server channels and realms and instead
delegates that responsibility to sysman. This way, sysman can act as the
bouncer for server connections.

Additionally, instead of launching vfsman (which then loads sysman),
the vfsman server now lives in sysman and sysman is the system init process.

Finally, add `eve_once` as a suspendable alternative for `sys_once` that
can safely be used in eve work items.